Refactor split out pulse_analyzer
