+++ src/CMakeLists.txt
@@ -20,6 +20,7 @@ add_library(r_433 STATIC
     optparse.c
     output_influx.c
     output_mqtt.c
+    pulse_analyzer.c
     pulse_demod.c
     pulse_detect.c
     pulse_detect_fsk.c
